backend/merchants/urls.py
# merchants/urls.py
from rest_framework.routers import DefaultRouter
from .views import MerchantViewSet

router = DefaultRouter()
router.register('merchants', MerchantViewSet)

urlpatterns = router.urls
